+++ src/lib/documents/New-CosmosDbDocument.ps1
@@ -52,7 +52,7 @@ function New-CosmosDbDocument
 
         [Parameter()]
         [ValidateNotNullOrEmpty()]
-        [System.String[]]
+        [System.Object[]]
         $PartitionKey,
 
         [Parameter()]
@@ -88,7 +88,7 @@ function New-CosmosDbDocument
     if ($PSBoundParameters.ContainsKey('PartitionKey'))
     {
         $headers += @{
-            'x-ms-documentdb-partitionkey' = '["' + ($PartitionKey -join '","') + '"]'
+            'x-ms-documentdb-partitionkey' = Format-CosmosDbDocumentPartitionKey -PartitionKey $PartitionKey
         }
         $null = $PSBoundParameters.Remove('PartitionKey')
     }