+++ test/Unit/CosmosDB.documents.Tests.ps1
@@ -151,6 +151,55 @@ InModuleScope CosmosDB {
         }
     }
 
+
+    Describe 'Format-CosmosDbDocumentPartitionKey' -Tag 'Unit' {
+        It 'Should exist' {
+            { Get-Command -Name Format-CosmosDbDocumentPartitionKey -ErrorAction Stop } | Should -Not -Throw
+        }
+
+        Context 'When called with a partition key with "abc"' {
+            It 'Should return ["abc"]' {
+                Format-CosmosDbDocumentPartitionKey -PartitionKey 'abc' | Should -Be '["abc"]'
+            }
+        }
+
+        Context 'When called with a partition key with "abc,def"' {
+            It 'Should return ["abc","def"]' {
+                Format-CosmosDbDocumentPartitionKey -PartitionKey 'abc','def' | Should -Be '["abc","def"]'
+            }
+        }
+
+        Context 'When called with a partition key with 123' {
+            It 'Should return [123]' {
+                Format-CosmosDbDocumentPartitionKey -PartitionKey 123 | Should -Be '[123]'
+            }
+        }
+
+        Context 'When called with a partition key with 123,456' {
+            It 'Should return [123,456]' {
+                Format-CosmosDbDocumentPartitionKey -PartitionKey 123,456 | Should -Be '[123,456]'
+            }
+        }
+
+        Context 'When called with a partition key with "abc",456' {
+            It 'Should return ["abc",456]' {
+                Format-CosmosDbDocumentPartitionKey -PartitionKey 'abc',456 | Should -Be '["abc",456]'
+            }
+        }
+
+        Context 'When called with a partition key with an unsupported type' {
+            It 'Should throw expected exception' {
+                $errorRecord = Get-InvalidArgumentRecord `
+                    -Message ($LocalizedData.ErrorPartitionKeyUnsupportedType -f 'True', 'System.Boolean') `
+                    -ArgumentName 'PartitionKey'
+
+                {
+                    Format-CosmosDbDocumentPartitionKey -PartitionKey $true
+                } | Should -Throw $errorRecord
+            }
+        }
+    }
+
     Describe 'Get-CosmosDbDocumentResourcePath' -Tag 'Unit' {
         It 'Should exist' {
             { Get-Command -Name Get-CosmosDbDocumentResourcePath -ErrorAction Stop } | Should -Not -Throw
@@ -164,6 +213,7 @@ InModuleScope CosmosDB {
                     Database     = $script:testDatabase
                     CollectionId = $script:testCollection
                     Id           = $script:testDocument1
+                    Verbose      = $true
                 }
 
                 { $script:result = Get-CosmosDbDocumentResourcePath @getCosmosDbDocumentResourcePathParameters } | Should -Not -Throw
@@ -200,6 +250,7 @@ InModuleScope CosmosDB {
                     ConsistencyLevel    = 'Strong'
                     SessionToken        = 'session'
                     PartitionKeyRangeId = 'partition'
+                    Verbose             = $true
                 }
 
                 { $script:result = Get-CosmosDbDocument @getCosmosDbDocumentParameters } | Should -Not -Throw
@@ -234,9 +285,10 @@ InModuleScope CosmosDB {
                 $script:ResponseHeader = $null
 
                 $getCosmosDbDocumentParameters = @{
-                    Context       = $script:testContext
-                    CollectionId  = $script:testCollection
+                    Context        = $script:testContext
+                    CollectionId   = $script:testCollection
                     ResponseHeader = [ref] $script:ResponseHeader
+                    Verbose        = $true
                 }
 
                 { $script:result = Get-CosmosDbDocument @getCosmosDbDocumentParameters } | Should -Not -Throw
@@ -274,6 +326,7 @@ InModuleScope CosmosDB {
                     Context      = $script:testContext
                     CollectionId = $script:testCollection
                     Id           = $script:testDocument1
+                    Verbose      = $true
                 }
 
                 { $script:result = Get-CosmosDbDocument @getCosmosDbDocumentParameters } | Should -Not -Throw
@@ -311,6 +364,7 @@ InModuleScope CosmosDB {
                     CollectionId = $script:testCollection
                     Id           = $script:testDocument1
                     PartitionKey = $script:testPartitionKey
+                    Verbose      = $true
                 }
 
                 { $script:result = Get-CosmosDbDocument @getCosmosDbDocumentParameters } | Should -Not -Throw
@@ -350,6 +404,7 @@ InModuleScope CosmosDB {
                     Context      = $script:testContext
                     CollectionId = $script:testCollection
                     DocumentBody = $script:testDocumentBody
+                    Verbose      = $true
                 }
 
                 { $script:result = New-CosmosDbDocument @newCosmosDbDocumentParameters } | Should -Not -Throw
@@ -371,8 +426,8 @@ InModuleScope CosmosDB {
             $script:result = $null
             $invokeCosmosDbRequest_parameterfilter = {
                 $Method -eq 'Post' -and `
-                $ResourceType -eq 'docs' -and `
-                $Encoding -eq 'UTF-8'
+                    $ResourceType -eq 'docs' -and `
+                    $Encoding -eq 'UTF-8'
             }
 
             Mock `
@@ -385,6 +440,7 @@ InModuleScope CosmosDB {
                     CollectionId = $script:testCollection
                     DocumentBody = $script:testDocumentBody
                     Encoding     = 'UTF-8'
+                    Verbose      = $true
                 }
 
                 { $script:result = New-CosmosDbDocument @newCosmosDbDocumentParameters } | Should -Not -Throw
@@ -420,6 +476,7 @@ InModuleScope CosmosDB {
                     CollectionId = $script:testCollection
                     DocumentBody = $script:testDocumentBody
                     PartitionKey = $script:testPartitionKey
+                    Verbose      = $true
                 }
 
                 { $script:result = New-CosmosDbDocument @newCosmosDbDocumentParameters } | Should -Not -Throw
@@ -459,6 +516,7 @@ InModuleScope CosmosDB {
                     Context      = $script:testContext
                     CollectionId = $script:testCollection
                     Id           = $script:testDocument1
+                    Verbose      = $true
                 }
 
                 { $script:result = Remove-CosmosDbDocument @removeCosmosDbDocumentParameters } | Should -Not -Throw
@@ -490,6 +548,7 @@ InModuleScope CosmosDB {
                     CollectionId = $script:testCollection
                     Id           = $script:testDocument1
                     PartitionKey = $script:testPartitionKey
+                    Verbose      = $true
                 }
 
                 { $script:result = Remove-CosmosDbDocument @removeCosmosDbDocumentParameters } | Should -Not -Throw
@@ -526,6 +585,7 @@ InModuleScope CosmosDB {
                     CollectionId = $script:testCollection
                     Id           = $script:testDocument1
                     DocumentBody = $script:testDocumentBody
+                    Verbose      = $true
                 }
 
                 { $script:result = Set-CosmosDbDocument @setCosmosDbDocumentParameters } | Should -Not -Throw
@@ -562,6 +622,7 @@ InModuleScope CosmosDB {
                     Id           = $script:testDocument1
                     DocumentBody = $script:testDocumentBody
                     PartitionKey = $script:testPartitionKey
+                    Verbose      = $true
                 }
 
                 { $script:result = Set-CosmosDbDocument @setCosmosDbDocumentParameters } | Should -Not -Throw
@@ -583,8 +644,8 @@ InModuleScope CosmosDB {
             $script:result = $null
             $invokeCosmosDbRequest_parameterfilter = {
                 $Method -eq 'Put' -and `
-                $ResourceType -eq 'docs' -and `
-                $Encoding -eq 'UTF-8'
+                    $ResourceType -eq 'docs' -and `
+                    $Encoding -eq 'UTF-8'
             }
 
             Mock `
@@ -598,6 +659,7 @@ InModuleScope CosmosDB {
                     Id           = $script:testDocument1
                     DocumentBody = $script:testDocumentBody
                     Encoding     = 'UTF-8'
+                    Verbose      = $true
                 }
 
                 { $script:result = Set-CosmosDbDocument @setCosmosDbDocumentParameters } | Should -Not -Throw
@@ -619,8 +681,8 @@ InModuleScope CosmosDB {
             $script:result = $null
             $invokeCosmosDbRequest_parameterfilter = {
                 $Method -eq 'Put' -and `
-                $ResourceType -eq 'docs' -and `
-                $Headers['If-Match'] -eq ($script:testETag)
+                    $ResourceType -eq 'docs' -and `
+                    $Headers['If-Match'] -eq ($script:testETag)
             }
 
             Mock `
@@ -634,6 +696,7 @@ InModuleScope CosmosDB {
                     Id           = $script:testDocument1
                     DocumentBody = $script:testDocumentBody
                     ETag         = $script:testETag
+                    Verbose      = $true
                 }
 
                 { $script:result = Set-CosmosDbDocument @setCosmosDbDocumentParameters } | Should -Not -Throw
@@ -672,6 +735,7 @@ InModuleScope CosmosDB {
                     DocumentBody = $script:testDocumentBody
                     PartitionKey = $script:testPartitionKey
                     ETag         = $script:testETag
+                    Verbose      = $true
                 }
 
                 { $script:result = Set-CosmosDbDocument @setCosmosDbDocumentParameters } | Should -Not -Throw