+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+
+- Updated `BuildHelpers` support module for CI pipelines to 2.0.10.
+
 ## 3.5.0.425
 
 - Fix issue with integration test execution error in Azure DevOps