+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## Unreleased
+
+- Change `psakefile.ps1` to detect Azure Pipelines correctly.
+- Updated `BuildHelpers` support module for CI pipelines to 2.0.10.
+- Added PowerShell Gallery badge to `README.md`.
+- Refactored `Invoke-CosmosDbRequest` and added
+  `Get-CosmosDbAuthorizationHeadersFromContext` to split out function to pull
+  token out of `Context` object. This was done to reduce the size of the
+  `Invoke-CosmosDbRequest` function and to improve testability.
+- Fix TravisCI MacOS build - fixes [Issue #313](https://github.com/PlagueHO/CosmosDB/issues/313).
+- Created helper function `Format-CosmosDbDocumentPartitionKey` to
+  format the partition key string passed to `x-ms-documentdb-partitionkey`
+  in document functions.
+- Add support for integer partition keys to functions
+  `Get-CosmosDbDocument`, `New-CosmosDbDocument`, `Remove-CosmosDbDocument`
+  and `Set-CosmosDbDocument`.
+
 ## 3.5.0.425
 
 - Fix issue with integration test execution error in Azure DevOps